Split class definitions out to separate files
