+++ src/main/kotlin/io/poyarzun/concoursedsl/domain/Resource.kt
@@ -5,8 +5,6 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategy
 import com.fasterxml.jackson.databind.annotation.JsonNaming
 import io.poyarzun.concoursedsl.dsl.DslList
 import io.poyarzun.concoursedsl.dsl.DslMap
-import io.poyarzun.concoursedsl.dsl.Tags
-import io.poyarzun.concoursedsl.dsl.Version
 
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy::class)
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
